feat: add new gateway-common crate
Part of introducing this create is moving the following pieces over to it from
the `graph-gateway` crate:

- chain-related types
- block constraints
- different error types
- metrics
