+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["graph-gateway", "indexer-selection", "prelude"]
+members = ["graph-gateway", "indexer-selection", "prelude", "common"]
 resolver = "2"
 
 [profile.release]
