+++ common/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+edition = "2021"
+name = "gateway-common"
+version = "0.0.1"
+
+[dependencies]
+alloy-primitives.workspace = true
+anyhow.workspace = true
+eventuals = "0.6.7"
+indexer-selection = { path = "../indexer-selection" }
+itertools = "0.12.0"
+lazy_static = "1.4.0"
+prelude = { path = "../prelude" }
+prometheus = "0.13.3"
+reqwest.workspace = true
+# rand.workspace = true
+# siphasher.workspace = true
+serde.workspace = true
+serde_json = "1.0.108"
+serde_with = "3.4.0"
+thegraph.workspace = true
+thiserror = "1.0.40"
+tokio.workspace = true
+toolshed.workspace = true
+tracing.workspace = true
