+++ graph-gateway/src/lib.rs
@@ -6,16 +6,13 @@ use reqwest::header::{self, HeaderMap, HeaderName, HeaderValue};
 pub mod auth;
 pub mod block_constraints;
 pub mod budgets;
-pub mod chains;
 pub mod client_query;
 pub mod config;
-pub mod errors;
 pub mod geoip;
 pub mod indexer_client;
 pub mod indexers;
 pub mod indexings_blocklist;
 pub mod ipfs;
-pub mod metrics;
 pub mod network_subgraph;
 pub mod receipts;
 pub mod reports;
